Fixed return type of uio_write and stdio_write.
git-svn-id: https://sc2.svn.sourceforge.net/svnroot/sc2/trunk/sc2@1079 8092fc87-c524-0410-9efc-e669fe64eaf9

+++ src/sc2code/libs/uio/io.h
@@ -107,7 +107,7 @@ int uio_rmdir(uio_DirHandle *dirHandle, const char *path);
 
 int uio_lseek(uio_Handle *handle, off_t offset, int whence);
 
-int uio_write(uio_Handle *handle, const void *buf, size_t count);
+ssize_t uio_write(uio_Handle *handle, const void *buf, size_t count);
 
 int uio_unlink(uio_DirHandle *dirHandle, const char *path);
